[resources] use to_bytes when converting routes
