+++ scrapyrt/log.py
@@ -5,20 +5,6 @@ import logging
 import os
 import sys
 
-try:
-    from scrapy.utils.python import to_bytes
-except ImportError:
-    def to_bytes(text, encoding=None, errors='strict'):
-        """Return the binary representation of `text`. If `text`
-        is already a bytes object, return it as-is."""
-        if isinstance(text, bytes):
-            return text
-        if not isinstance(text, six.string_types):
-            raise TypeError('to_bytes must receive a unicode, str or bytes '
-                            'object, got %s' % type(text).__name__)
-        if encoding is None:
-            encoding = 'utf-8'
-        return text.encode(encoding, errors)
 
 from twisted.python import log
 from twisted.python.log import startLoggingWithObserver
@@ -26,6 +12,7 @@ from twisted.python.logfile import DailyLogFile
 from scrapy.settings import Settings
 from scrapy.utils.log import DEFAULT_LOGGING, TopLevelFormatter
 
+from scrapyrt.utils import to_bytes
 from .conf import settings as scrapyrt_settings
 
 DEBUG = logging.DEBUG
