+++ scrapyrt/utils.py
@@ -25,3 +25,19 @@ def extract_scrapy_request_args(dictionary, raise_error=False):
 
 def is_python2():
     return sys.version_info < (3,0)
+
+
+try:
+    from scrapy.utils.python import to_bytes
+except ImportError:
+    def to_bytes(text, encoding=None, errors='strict'):
+        """Return the binary representation of `text`. If `text`
+        is already a bytes object, return it as-is."""
+        if isinstance(text, bytes):
+            return text
+        if not isinstance(text, six.string_types):
+            raise TypeError('to_bytes must receive a unicode, str or bytes '
+                            'object, got %s' % type(text).__name__)
+        if encoding is None:
+            encoding = 'utf-8'
+        return text.encode(encoding, errors)
